Convert Effect to be generic
